Add @Before e @After para contabilizar testes

+++ src/FontesCalculadora/CalculadoraAritmetica.java
@@ -2,19 +2,19 @@ package FontesCalculadora;
 
 public class CalculadoraAritmetica {
 	
-	public int dividir(int num1, int num2) {
+	public double dividir(double num1, double num2) {
 		return num1 / num2;
 	}
 
-	public int multiplicar(int num1, int num2) {
+	public double multiplicar(double num1, double num2) {
 		return num1 * num2;
 	}
 
-	public int subtrair(int num1, int num2) {
+	public double subtrair(double num1, double num2) {
 		return num1 - num2;
 	}
 
-	public int soma (int num1, int num2) {
+	public double soma (double num1, double num2) {
 		return num1 + num2;
 	}
 }
